1.0.9

+++ Includes/Modbus_Slave.h
@@ -1,9 +1,11 @@
 
-// Author    KMS - Martin Dubois, P. Eng.
-// Copyright (C) 2024 KMS
-// License   http://www.apache.org/licenses/LICENSE-2.0
 // Product   KMS-uC
-// File      Includes/Modbus_Slave.h
+// License   http://www.apache.org/licenses/LICENSE-2.0
+
+/// \author    KMS - Martin Dubois, P. Eng.
+/// \copyright Copyright &copy; 2024-2026 KMS
+/// \file      Includes/Modbus_Slave.h
+/// \brief     Functions to process Modbus commands
 
 #pragma once
 
@@ -15,9 +17,14 @@
 
 struct Modbus_Slave_Range_s;
 
-// Return  MODBUS_NO_ERROR
-//         MODBUS_EXCEPTION_...
-typedef uint8_t (*Modbus_Slave_Callback)(struct Modbus_Slave_Range_s* aRange, uint16_t aAddress, uint8_t aCount, uint16_t* aData);
+/// \brief Modbus callback
+/// \param aRange   The address range
+/// \param aAddress Start address
+/// \param aCount   Register count
+/// \param aData    Data
+/// \retval MODBUS_NO_ERROR
+/// \retval MODBUS_EXCEPTION_...
+typedef uint8_t (*Modbus_Slave_Callback)(struct Modbus_Slave_Range_s* aRange, uint16_t aAddress, uint16_t aCount, uint16_t* aData);
 
 // mContext      Way to pass data to the callbacks
 // mAddress      The starting address in the Modbus address space
@@ -33,6 +40,9 @@ typedef uint8_t (*Modbus_Slave_Callback)(struct Modbus_Slave_Range_s* aRange, ui
 // mBeforeWrite  Mandatory. If not needed, set it to
 //               Modbus_Slave_Callback_Default. If the range is read only,
 //               set it to Modbus_Slave_Callback_Error
+
+/// \brief Modbus address rance
+/// \see Modbus_Slave_Callback Modbus_Slave_Init
 typedef struct Modbus_Slave_Range_s
 {
     void* mContext;
@@ -51,28 +61,45 @@ Modbus_Slave_Range;
 // Functions
 // //////////////////////////////////////////////////////////////////////////
 
-// aUART          The UART index
-// aDevice        The Modbus device address
-// aRanges        The register ranges
-// aRangeQty      The number of ranges
-// aOutputEnable  This GPIO to set to 1 when transmiting.
-//                    .mBit
-//                    .mDrive
-//                    .mInterrupt_Falling
-//                    .mOutput            : Ignored, must be an output
-//                    .mPort              : See GPIO_PORT_...
-//                    .mPull_Enable
-//                    .mPullUp_Select
-//                    .mPushPull
-//                    .mSlewRate_Slow     : Ignored, must be set
+/// \brief Initialize the module
+/// \param aUART         The UART index
+/// \param aDevice       The Modbus device address
+/// \param aRanges       The register ranges
+/// \param aRangeQty     The number of ranges
+/// \param aOutputEnable This GPIO to set to 1 when transmiting.
+
+// .mBit
+// .mDrive
+// .mInterrupt_Falling
+// .mOutput            : Ignored, must be an output
+// .mPort              : See GPIO_PORT_...
+// .mPull_Enable
+// .mPullUp_Select
+// .mPushPull
+// .mSlewRate_Slow     : Ignored, must be set
 extern void Modbus_Slave_Init(uint8_t aUART, uint8_t aDevice, Modbus_Slave_Range* aRanges, uint8_t aRangeQty, GPIO aOutputEnable);
 
-// Return  MODBUS_NO_ERROR
-extern uint8_t Modbus_Slave_Callback_Default(struct Modbus_Slave_Range_s* aRange, uint16_t aAddress, uint8_t aCount, uint16_t* aData);
+/// \brief Default callback
+/// \param aRange   The address range
+/// \param aAddress Start address
+/// \param aCount   Register count
+/// \param aData    Data
+/// \retval MODBUS_NO_ERROR
+/// \see Modbus_Slave_Callback
+extern uint8_t Modbus_Slave_Callback_Default(struct Modbus_Slave_Range_s* aRange, uint16_t aAddress, uint16_t aCount, uint16_t* aData);
 
-// Return  MODBUS_EXCEPTION_ILLEGAL_DATA_ADDRESS
-extern uint8_t Modbus_Slave_Callback_Error(struct Modbus_Slave_Range_s* aRange, uint16_t aAddress, uint8_t aCount, uint16_t* aData);
+/// \brief Default callback
+/// \param aRange   The address range
+/// \param aAddress Start address
+/// \param aCount   Register count
+/// \param aData    Data
+/// \retval MODBUS_EXCEPTION_ILLEGAL_DATA_ADDRESS
+/// \see Modbus_Slave_Callback
+extern uint8_t Modbus_Slave_Callback_Error(struct Modbus_Slave_Range_s* aRange, uint16_t aAddress, uint16_t aCount, uint16_t* aData);
 
+/// \brief Periodic work
+/// \param aPeriod_ms Delay since the last call
 extern void Modbus_Slave_Tick(uint16_t aPeriod_ms);
 
+/// \brief Idle work
 extern void Modbus_Slave_Work();
